+++ src/network.lisp
@@ -5,6 +5,8 @@
                 #:obfuscate)
   (:import-from #:cl-telegram-bot/bot
                 #:get-endpoint)
+  (:import-from #:trivial-timeout
+                #:with-timeout)
   (:export
    #:make-request
    #:request-error))
@@ -17,17 +19,21 @@
              (format stream "Request error: ~A" (what condition)))))
 
 
-(defun make-request (bot name options &key (streamp nil) timeout)
+(defun make-request (bot name options &key (streamp nil) (timeout 3))
   "Perform HTTP request to 'name API method with 'options JSON-encoded object."
   (let ((url (concatenate 'string (get-endpoint bot) name)))
     (log:debug "Posting data to"
                (obfuscate url)
                options)
-    (let* ((response (dexador:post url
-                                   :stream streamp
-                                   :headers '(("Content-Type" . "application/json"))
-                                   :content (jonathan:to-json options)
-                                   :timeout timeout))
+    (let* ((max-timeout (* timeout 10))
+           (response
+             ;; TODO: Try new dexador with connect-timeout and read-timeout
+             (with-timeout (max-timeout)
+               (dexador:post url
+                             :stream streamp
+                             :headers '(("Content-Type" . "application/json"))
+                             :content (jonathan:to-json options)
+                             :timeout timeout)))
            (data (jonathan:parse response)))
       (unless (getf data :|ok|)
         (log:error "Wrong data received from the server" data)